requirements.txt
requests~=2.31.0
pillow~=9.4.0
requests